refactor : user entity, repository 위치 변경
