+++ build.gradle
@@ -6,6 +6,7 @@ plugins {
 
 group = 'com.newspeed19'
 version = '0.0.1-SNAPSHOT'
+
 sourceCompatibility = '17'
 
 repositories {
