+++ src/main/java/com/newspeed19/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.newspeed19.profile.repository;
+package com.newspeed19.user.repository;
 
-import com.newspeed19.common.User;
+import com.newspeed19.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
